+++ packages/gestalt/src/Datapoint/InternalDatapoint.js
@@ -1,5 +1,6 @@
 // @flow strict
 import { type Node } from 'react';
+import AccessibilityPause from '../accessibility/AccessibilityPause.js';
 import Badge from '../Badge.js';
 import Box from '../Box.js';
 import Flex from '../Flex.js';
@@ -67,6 +68,7 @@ export default function InternalDatapoint({
         <MaybeMinWidth minWidth={minTitleWidth}>
           <Text size="200" color={textColor} lineClamp={lineClamp}>
             {title}
+            <AccessibilityPause />
           </Text>
         </MaybeMinWidth>
         {tooltipText && (
