+++ packages/gestalt/src/accessibility/AccessibilityPause.js
@@ -1,0 +1,7 @@
+// @flow strict
+import { type Node } from 'react';
+import VisuallyHidden from './VisuallyHidden.js';
+
+export default function AccessibilityPause(): Node {
+  return <VisuallyHidden>,</VisuallyHidden>;
+}
